+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 0.9.6 (2021-10-05)
 ------------------
 
+- Drop support for Python 2 #61
 - Add support for new github URLs in url2purl #47
 
 0.9.5 (2021-10-04)